Mcp tools refactor
