+++ mcp_tools/web_scan/nikto.py
@@ -1,0 +1,28 @@
+# mcp_tools/web_scan/nikto.py
+
+from typing import Dict, Any
+
+def register_nikto_tool(mcp, hexstrike_client, logger):
+    @mcp.tool()
+    def nikto_scan(target: str, additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute Nikto web vulnerability scanner with enhanced logging.
+
+        Args:
+            target: The target URL or IP
+            additional_args: Additional Nikto arguments
+
+        Returns:
+            Scan results with discovered vulnerabilities
+        """
+        data = {
+            "target": target,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔬 Starting Nikto scan: {target}")
+        result = hexstrike_client.safe_post("api/tools/nikto", data)
+        if result.get("success"):
+            logger.info(f"✅ Nikto scan completed for {target}")
+        else:
+            logger.error(f"❌ Nikto scan failed for {target}")
+        return result
